+++ controllers/v1/programs.js
@@ -396,20 +396,17 @@ module.exports = class Programs extends Abstract {
     * {
         "entities": {
             "district": [
-                "ALL"
+                "682303044e2812081f3426fb"
             ],
             "professional_subroles": [
-                "ALL",
                 "682301604e2812081f342674",
                 "682303044e2812081f3426fb"
             ],
             "professional_role": [
-                "ALL",
                 "681b07b49c57cdcf03c79ae3",
                 "681b0800f21c88cef9517e0e"
             ],
             "school": [
-                "ALL",
                 "67c82d9553812588916410d3"
             ],
             "language": [
@@ -536,11 +533,9 @@ module.exports = class Programs extends Abstract {
             "professional_subroles": [
                 "682301254e2812081f34266c",
                 "682303044e2812081f3426fb",
-                "ALL",
                 "682301604e2812081f342674"
             ],
             "professional_role": [
-                "ALL",
                 "681b07b49c57cdcf03c79ae3",
                 "681b0800f21c88cef9517e0e"
             ]
